feat(GoogleMerchant): Allow to load products via scope.
Currently all products are loaded from the database during
application boot. This is not ideal, especially if you
want to boot the app to prepare specs or build assets on
heroku or in a docker container. Let's agree that accessing the
database during app boot is generally a bad idea.

You can now pass a `product_scope` instead of passing products.
Default scope is `available` (which is what most stores want anyway).

So, it now gets as simple as

    SolidusFeeds::Generators::GoogleMerchant.new(
      host: "www.example.com", # You should pass this instead of using
      # the Spree::Store.current, which as well accesses the database
    )

or

    SolidusFeeds::Generators::GoogleMerchant.new(
      product_scope: :solidus_mugs,
      host: "www.example.com"
    )